+++ runtime/src/main/3.3.2-scala/org/apache/spark/sql/ShimUtils.scala
@@ -342,4 +342,5 @@ object ShimUtils {
    */
   def isUsable(sparkSession: SparkSession): Boolean =
     !sparkSession.sparkContext.isStopped
+
 }